+++ crates/rslint_core/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Moved util from a file to its own directory
+- Removed the `module` parameter from `lint_file` and replaced it with `syntax: Syntax`
 
 ## [0.2.1] - 2020-10-21